Move get_yaml_data_structure method from Loader to DataStructureLoader class
