+++ test/test_schema_data_structure.py
@@ -7,7 +7,7 @@ import pytest
 from ve_utils.utype import UType as Ut
 from jsonschema.exceptions import SchemaError
 from vemonitor_m8.confManager.schema_validate import SchemaValidate
-from vemonitor_m8.confManager.loader import Loader
+from vemonitor_m8.confManager.data_structure_loader import DataStructureLoader
 from .schema_test_helper import SchemaTestHelper
 
 @pytest.fixture(name="schema_manager", scope="class")
@@ -18,8 +18,12 @@ def schema_manager_fixture():
         def __init__(self):
             SchemaTestHelper.__init__(self)
             self.schema = SchemaValidate.load_schema("data_structure")
-            loader = Loader("vemonitor_m8/confManager/confFiles/victronDeviceData.yaml")
-            self.obj = loader.get_yaml_data_structure(file_path="victronDeviceData.yaml")
+            loader = DataStructureLoader(
+                "vemonitor_m8/confManager/confFiles/victronDeviceData.yaml"
+            )
+            self.obj = loader.get_yaml_data_structure(
+                file_path="victronDeviceData.yaml"
+            )
 
     return SchemaManager()
